change count data type to int

+++ psifr/transitions.py
@@ -161,8 +161,8 @@ def count_pairs(n_item, pool_items, recall_items,
                 pool_test=None, recall_test=None, test=None):
     """Count transitions between pairs of specific items."""
 
-    actual = np.zeros((n_item, n_item))
-    possible = np.zeros((n_item, n_item))
+    actual = np.zeros((n_item, n_item), dtype=int)
+    possible = np.zeros((n_item, n_item), dtype=int)
     for i, recall_items_list in enumerate(recall_items):
         # set up masker to filter transitions
         pool_test_list = None if pool_test is None else pool_test[i]
